pyproject.toml
[project]
name = "weather-app"
version = "0.1.0"
description = "A simple weather application that provides current weather information based on your location."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
    "streamlit-geolocation>=0.0.10",
]
